scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = "A scraper for Internshala internship listings, designed for CRON job execution."
authors = ["Sarthak Kimtani <sarthakkimtani123@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
sqlalchemy = "^1.4.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^0.19.0"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
